changing maxlen from len to size_t http://stackoverflow.com/q/31720747/719016 -- adding runfiles -- adding deinterleave.c

+++ src/mystring.h
@@ -24,7 +24,7 @@
 
 struct __mystring_struct {
    char * string;
-   int len, maxlen;
+   size_t len, maxlen;
 };
 typedef struct __mystring_struct * Mystring;
 #define Mystring_size sizeof(struct __mystring_struct)
